Explicitly specify 16 bit ints for audio output bits and analog readings.
This does not make any difference on AVR, but for 32bit boards such as
Teensy or STM32, "int" is actually 32bits. While this is not much of a
problem in most places, it causes excessive memory consumption in the
audio output buffer (and to a lesser degree in the analog readings buffer).

Unfortunately this will break existing sketches on 32 bit boards (but only
on these), as it changes the signature of updateAudio().