Call updateControl() from audioHook(), instead of timer - Part1
This is part one of the patch, removing all uses of Timer 0, and, instead,
inserting code to call updateControl() from inside audioHook() at the
specified rate. NOTE that the utility/TimerZero-header (and source) is _not_
removed, just in case users have started using it, or might find it of
interest, now that Timer 0 is free for use.