Call updateControl() from audioHook(), instead of timer - Part2
This is part two of the patch, removing all uses of ATOMIC_BLOCK, since
there is no longer any danger of entering one of these functions from the
control interrupt, while they are in the middle of executing in the main
loop.

Should this cause any unexepcted problems, this part of the patch can be
reverted independently of part one (but not the other way around).