Allow to skip ADC setup.
E.g. for compatibility with certain libraries, or to enable techniques such as
analog port multiplexing.

Note that this is a runtime solution. On the positive side, users can select to
skip ADC setup per sketch, without making any changes in the lib. On the negative side
(compared to a compile-time switch), this solution does not save any allocated RAM, and
the ISR routine is still defined on platforms that define it.