Polish the proposed API some more. Importantly, provide the illusion of a struct return value in mono sketches.
This is just syntactic sugar that the compiler will burn through (on both -Os, the Arduino default, and -O1/-O2),
but allows several convenience functions, such as a clip() function and a compatible API for mono and stereo samples.

I'll still have to re-adjust the examples for this. But also with this revision, existing sketeches continue to work
unmodified.