Add a multi-complation-unit example
This is pretty pointless at the time being, but will be marginally useful, to demonstrate how to customize config
in that case, once Mozzi itself is compiled as single compilation unit.

Add it now, already (at the Mozzi2-branch), so it is available as a baseline for the "report" action in future PRs.