Remove STEREO_HACK
This has been deprecated for several years, now. Also, as this is a config option, projects using it will need to adjust, anyway, so it
does not actually help to keep sketches working.

While in the process, turn AudioOutput and AudioOutput_t into typedefs, rather than #defines to reduce the macro-mess. This required
moving some bits around, however.