CONTROL_RATE -> MOZZI_CONTROL_RATE
While the old name continues to work, let's be consistent in the examples.
For the same reason (and to silence some legit compiler warnings; consider what happends when changing after inclusion
of PDResonsant.h, for example!), move it above #include <Mozzi.h>, in all examples.

Change occurences of startMozzi(CONTROL_RATE) to simply startMozzi(), which does the same thing.